fix(pe): add auth header to callbacks + fix check_budget return type
- CallbackService now accepts api_key param and sends x-api-key header,
  fixing 401 when PE posts results back to NC's /webhooks/processing-done
- Worker passes Settings().api_key to CallbackService
- check_budget SQL function: add ::REAL cast to pct_used column,
  fixing asyncpg DatatypeMismatchError (double precision vs real)

+++ processing-engine/alembic/versions/010_create_functions.py
@@ -76,7 +76,7 @@ def upgrade() -> None:
               AND l.created_at >= v_start
               AND l.status = 'success';
 
-            RETURN QUERY SELECT v_cost, v_budget, (v_cost / v_budget * 100.0), (v_cost >= v_budget);
+            RETURN QUERY SELECT v_cost, v_budget, (v_cost / v_budget * 100.0)::REAL, (v_cost >= v_budget);
         END;
         $$ LANGUAGE plpgsql STABLE
     """)
